R.KEYWORDS {R.methodsS3}	R Documentation

Reserved words in R not to be used for object names

Description

Reserved words in R not to be used for object names. R.KEYWORDS is a character vector of all reserved words in R according to [1].

Author(s)

Henrik Bengtsson

References

[1] Section "Reserved words", R Language Definition, version 2.6.0 (2007-09-14) DRAFT.

[Package R.methodsS3 version 1.7.1 Index]