getDispatchMethodS3 {R.methodsS3}	R Documentation

Gets the S3 method that a generic function would call

Description

Gets the S3 method that a generic function would call according to an S3 class() vector.

Usage

## Default S3 method:
getDispatchMethodS3(methodName, classNames, ...)

Arguments

methodName
A character string specifying the name of a generic function.

classNames
A character vector of class() names.

...
Not used.

Value

Returns a function, or throws an exception if not found.

Author(s)

Henrik Bengtsson

See Also

findDispatchMethodsS3().

[Package R.methodsS3 version 1.8.2 Index]