getGenericS3 {R.methodsS3}	R Documentation

Gets an S3 generic function

Description

Gets an S3 generic function.

Usage

## Default S3 method:
getGenericS3(name, envir=parent.frame(), inherits=TRUE, ...)

Arguments

name
The name of the generic function.

envir
The environment from which the search for the generic function is done.

inherits
A logical specifying whether the enclosing frames should be searched or not.

...
Not used.

Author(s)

Henrik Bengtsson

See Also

setGenericS3(). getMethodS3(). isGenericS3().

[Package R.methodsS3 version 1.8.2 Index]