isGenericS3 {R.methodsS3}	R Documentation

Checks if a function is a S3 generic function

Description

Checks if a function is a S3 generic function.

Usage

## Default S3 method:
isGenericS3(fcn, envir=parent.frame(), ...)

Arguments

fcn
A function or a character string.

envir
If argument fcn is a character, this is the environment from which the search for the function is done.

...
Not used.

Details

A function is considered to be a generic S3/UseMethod function if its name matches one of the known S3 generic functions, or if it calls UseMethod().

Value

Returns TRUE if a generic S3/UseMethod function, otherwise FALSE.

Author(s)

Henrik Bengtsson

[Package R.methodsS3 version 1.8.2 Index]