findDispatchMethodsS3 {R.methodsS3}	R Documentation

Finds the S3 methods that a generic function would call

Description

Finds the S3 methods that a generic function would call, ordered according to an S3 class() vector.

Usage

## Default S3 method:
findDispatchMethodsS3(methodName, classNames, firstOnly=FALSE, ...)

Arguments

methodName
A character string specifying the name of a generic function.

classNames
A character vector of class() names.

firstOnly
If TRUE, only the first method is returned.

...
Not used.

Value

Returns a names list structure.

Author(s)

Henrik Bengtsson

See Also

getDispatchMethodS3().

[Package R.methodsS3 version 1.7.1 Index]